* align OptimizedQueryParser and OptimizedMultiFieldQueryParser.
